refactor: split mcp, setup, tool-adapters, team, scope, generate, and embeddings into focused modules
- mcp/index.ts (847→260): extract formatters.ts, tools.ts, streaming-tools.ts
- setup.ts (415→210): extract setup-tools.ts with Tool interface and TOOLS data
- tool-adapters.ts (461→300): extract tool-adapters-helpers.ts with pure helpers
- team-content.tsx (1204→1060): extract team-types.ts and team-helpers.ts
- scope.ts (567→200): extract scope-parsers.ts and scope-schema.ts
- generate.ts (461→420): extract generate-targets.ts with Target data
- embeddings.ts (458→390): extract embeddings-models.ts with model registry

Adds 58 tests across 8 new test files covering all extracted modules.
