+++ packages/cli/src/commands/generate.ts
@@ -14,74 +14,10 @@ import {
   runAdapter,
   logAdaptResult,
 } from "../lib/tool-adapters.js";
-import { formatMemoriesAsMarkdown, formatCursorMdc, formatWindsurf } from "../lib/formatters.js";
-import { MARKER, makeFooter, hasOurMarker } from "../lib/markers.js";
-
-// ── Target Registry ─────────────────────────────────────────────────
-
-interface Target {
-  name: string;
-  defaultPath: string;
-  description: string;
-  format: (memories: Memory[]) => string;
-}
-
-const TARGETS: Target[] = [
-  {
-    name: "cursor",
-    defaultPath: ".cursor/rules/memories.mdc",
-    description: "Cursor rules file (.cursor/rules/memories.mdc)",
-    format: formatCursorMdc,
-  },
-  {
-    name: "claude",
-    defaultPath: "CLAUDE.md",
-    description: "Claude Code instructions (CLAUDE.md)",
-    format: (m) => `# Project Memories\n\n${formatMemoriesAsMarkdown(m)}`,
-  },
-  {
-    name: "agents",
-    defaultPath: ".agents/",
-    description: ".agents/ directory (instructions, rules, skills, settings)",
-    // format is unused for agents — handled by generateAgentsDir
-    format: (m) => `# Project Memories\n\n${formatMemoriesAsMarkdown(m)}`,
-  },
-  {
-    name: "copilot",
-    defaultPath: ".github/copilot-instructions.md",
-    description: "GitHub Copilot instructions (.github/copilot-instructions.md)",
-    format: (m) => `# Project Memories\n\n${formatMemoriesAsMarkdown(m)}`,
-  },
-  {
-    name: "windsurf",
-    defaultPath: ".windsurf/rules/memories.md",
-    description: "Windsurf rules (.windsurf/rules/memories.md)",
-    format: formatWindsurf,
-  },
-  {
-    name: "cline",
-    defaultPath: ".clinerules/memories.md",
-    description: "Cline rules (.clinerules/memories.md)",
-    format: (m) => `# Project Memories\n\n${formatMemoriesAsMarkdown(m)}`,
-  },
-  {
-    name: "roo",
-    defaultPath: ".roo/rules/memories.md",
-    description: "Roo rules (.roo/rules/memories.md)",
-    format: (m) => `# Project Memories\n\n${formatMemoriesAsMarkdown(m)}`,
-  },
-  {
-    name: "gemini",
-    defaultPath: "GEMINI.md",
-    description: "Gemini instructions (GEMINI.md)",
-    format: (m) => `# Project Memories\n\n${formatMemoriesAsMarkdown(m)}`,
-  },
-];
+import { makeFooter, hasOurMarker } from "../lib/markers.js";
+import { type Target, TARGETS, TRACK_BY_DEFAULT } from "./generate-targets.js";
 
 // ── .gitignore Management ───────────────────────────────────────────
-
-// Files users likely want git-tracked (shared with team)
-const TRACK_BY_DEFAULT = new Set(["CLAUDE.md", ".agents/", "GEMINI.md", ".github/copilot-instructions.md"]);
 
 async function checkGitignore(filePath: string): Promise<void> {
   // Only suggest for files in hidden directories (tool-specific)
