+++ packages/cli/src/lib/setup.ts
@@ -3,218 +3,15 @@ import { readFile, writeFile, mkdir } from "node:fs/promises";
 import { join, dirname } from "node:path";
 import { homedir } from "node:os";
 import { execSync } from "node:child_process";
-export interface Tool {
-  name: string;
-  detectPaths: string[];
-  detectCommands?: string[];
-  globalDetectPaths?: string[];
-  mcpConfigPath?: string;
-  globalMcpConfigPath?: string;
-  mcpConfigFormat?: "cursor" | "claude" | "vscode" | "opencode";
-  instructionFile?: string;
-  generateCmd?: string;
-  generateArgs?: string[];
-  setupHint?: string;
-}
+import {
+  type Tool,
+  TOOLS,
+  MEMORIES_MCP_CONFIG,
+  MEMORIES_MCP_CONFIG_OPENCODE,
+} from "./setup-tools.js";
 
-// Tool definitions with their config locations
-const TOOLS: Tool[] = [
-  {
-    name: "Cursor",
-    detectPaths: [".cursor", ".cursor/rules/memories.mdc"],
-    globalDetectPaths: [".cursor"],
-    mcpConfigPath: ".cursor/mcp.json",
-    mcpConfigFormat: "cursor",
-    instructionFile: ".cursor/rules/memories.mdc",
-    generateCmd: "cursor",
-  },
-  {
-    name: "Claude Code",
-    detectPaths: [".claude", "CLAUDE.md", ".mcp.json"],
-    globalDetectPaths: [".claude"],
-    mcpConfigPath: ".mcp.json",
-    mcpConfigFormat: "claude",
-    instructionFile: "CLAUDE.md",
-    generateCmd: "claude",
-  },
-  {
-    name: "Windsurf",
-    detectPaths: [".windsurf", ".windsurf/rules/memories.md", ".windsurfrules"],
-    globalDetectPaths: [".windsurf"],
-    mcpConfigPath: ".windsurf/mcp.json",
-    mcpConfigFormat: "cursor",
-    instructionFile: ".windsurf/rules/memories.md",
-    generateCmd: "windsurf",
-  },
-  {
-    name: "VS Code",
-    detectPaths: [".vscode", ".vscode/mcp.json"],
-    globalDetectPaths: [".vscode"],
-    mcpConfigPath: ".vscode/mcp.json",
-    mcpConfigFormat: "vscode",
-  },
-  {
-    name: "OpenCode",
-    detectPaths: [".opencode", ".opencode/instructions.md", "opencode.json", ".opencode/mcp.json"],
-    detectCommands: ["opencode"],
-    globalDetectPaths: [".config/opencode/opencode.json", ".opencode"],
-    mcpConfigPath: "opencode.json",
-    globalMcpConfigPath: ".config/opencode/opencode.json",
-    mcpConfigFormat: "opencode",
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-    setupHint: "OpenCode supports both .agents/ and opencode.json MCP configuration.",
-  },
-  {
-    name: "Factory",
-    detectPaths: [".factory", ".factory/instructions.md", ".factory/mcp.json"],
-    globalDetectPaths: [".factory"],
-    mcpConfigPath: ".factory/mcp.json",
-    mcpConfigFormat: "cursor",
-    instructionFile: ".factory/instructions.md",
-    generateCmd: "claude",
-    generateArgs: ["--output", ".factory/instructions.md"],
-  },
-  {
-    name: "Kiro",
-    detectPaths: [".kiro", ".kiro/settings/mcp.json", ".kiro/mcp.json"],
-    detectCommands: ["kiro"],
-    globalDetectPaths: [".kiro", ".kiro/settings/mcp.json"],
-    mcpConfigPath: ".kiro/settings/mcp.json",
-    mcpConfigFormat: "cursor",
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-  },
-  {
-    name: "Kilo",
-    detectPaths: [".kilo", ".kilo/mcp.json"],
-    detectCommands: ["kilo"],
-    globalDetectPaths: [".kilo", ".kilo/mcp.json"],
-    mcpConfigPath: ".kilo/mcp.json",
-    mcpConfigFormat: "cursor",
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-  },
-  {
-    name: "Trae",
-    detectPaths: [".trae", ".trae/mcp.json"],
-    detectCommands: ["trae"],
-    globalDetectPaths: [".trae", ".trae/mcp.json"],
-    mcpConfigPath: ".trae/mcp.json",
-    mcpConfigFormat: "cursor",
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-  },
-  {
-    name: "Antigravity",
-    detectPaths: [".antigravity", ".antigravity/mcp.json", "mcp_config.json"],
-    detectCommands: ["antigravity"],
-    globalDetectPaths: [".gemini/antigravity/mcp_config.json", ".antigravity"],
-    mcpConfigPath: ".antigravity/mcp.json",
-    globalMcpConfigPath: ".gemini/antigravity/mcp_config.json",
-    mcpConfigFormat: "cursor",
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-  },
-  {
-    name: "Goose",
-    detectPaths: [".goose", ".goose/rules"],
-    detectCommands: ["goose"],
-    globalDetectPaths: [".goose"],
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-    setupHint: "Configure Goose MCP in Goose settings if you need live memory access.",
-  },
-  {
-    name: "OpenClaw",
-    detectPaths: [
-      ".openclaw/workspace/AGENTS.md",
-      ".openclaw/workspace/SOUL.md",
-      ".openclaw/workspace/TOOLS.md",
-      ".openclaw/workspace/IDENTITY.md",
-      ".openclaw/workspace/USER.md",
-      ".openclaw/workspace/HEARTBEAT.md",
-      ".openclaw/workspace/BOOTSTRAP.md",
-      ".openclaw/workspace/BOOT.md",
-      ".openclaw/workspace/MEMORY.md",
-      ".openclaw/workspace/skills",
-      ".openclaw/workspace/memory",
-    ],
-    detectCommands: ["openclaw"],
-    globalDetectPaths: [
-      ".openclaw/workspace/AGENTS.md",
-      ".openclaw/workspace/SOUL.md",
-      ".openclaw/workspace/TOOLS.md",
-      ".openclaw/workspace/IDENTITY.md",
-      ".openclaw/workspace/USER.md",
-      ".openclaw/workspace/HEARTBEAT.md",
-      ".openclaw/workspace/BOOTSTRAP.md",
-      ".openclaw/workspace/BOOT.md",
-      ".openclaw/workspace/MEMORY.md",
-      ".openclaw/workspace/skills",
-      ".openclaw/workspace/memory",
-    ],
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-    setupHint: "OpenClaw uses workspace artifacts (AGENTS/SOUL/TOOLS/IDENTITY/USER/HEARTBEAT/BOOTSTRAP + skills) under ~/.openclaw/workspace.",
-  },
-  {
-    name: "Blackbox CLI",
-    detectPaths: [],
-    detectCommands: ["blackbox"],
-    setupHint: "Add memories manually: blackbox mcp add memories http://127.0.0.1:3030/mcp -t http",
-  },
-  {
-    name: "GitHub Copilot",
-    detectPaths: [".github/copilot-instructions.md"],
-    instructionFile: ".github/copilot-instructions.md",
-    generateCmd: "copilot",
-  },
-  {
-    name: "Gemini",
-    detectPaths: ["GEMINI.md"],
-    instructionFile: "GEMINI.md",
-    generateCmd: "gemini",
-  },
-  {
-    name: "Cline",
-    detectPaths: [".clinerules", ".clinerules/memories.md"],
-    instructionFile: ".clinerules/memories.md",
-    generateCmd: "cline",
-  },
-  {
-    name: "Roo",
-    detectPaths: [".roo", ".roo/rules/memories.md"],
-    instructionFile: ".roo/rules/memories.md",
-    generateCmd: "roo",
-  },
-  {
-    name: "Amp",
-    detectPaths: [".amp", ".amp/mcp.json"],
-    globalDetectPaths: [".amp"],
-    mcpConfigPath: ".amp/mcp.json",
-    mcpConfigFormat: "cursor",
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-  },
-  {
-    name: "Agent Harness (.agents)",
-    detectPaths: [".agents", ".agents/instructions.md", "AGENTS.md", ".codex", ".opencode"],
-    instructionFile: ".agents/instructions.md",
-    generateCmd: "agents",
-  },
-];
-
-// MCP server configuration for memories
-const MEMORIES_MCP_CONFIG = {
-  command: "npx",
-  args: ["-y", "@memories.sh/cli", "serve"],
-};
-
-const MEMORIES_MCP_CONFIG_OPENCODE = {
-  type: "local" as const,
-  command: ["npx", "-y", "@memories.sh/cli", "serve"],
-};
+// Re-export for backward compatibility
+export { type Tool, getAllTools } from "./setup-tools.js";
 
 interface McpConfig {
   mcpServers?: Record<string, { command: string; args: string[]; env?: Record<string, string> }>;
@@ -256,7 +53,7 @@ function commandExists(command: string): boolean {
   }
 }
 
-function parseConfigFile(content: string): McpConfig {
+export function parseConfigFile(content: string): McpConfig {
   try {
     return JSON.parse(content) as McpConfig;
   } catch {
@@ -265,13 +62,6 @@ function parseConfigFile(content: string): McpConfig {
     const withoutLineComments = withoutBlockComments.replace(/^\s*\/\/.*$/gm, "");
     return JSON.parse(withoutLineComments) as McpConfig;
   }
-}
-
-/**
- * Get all supported tools
- */
-export function getAllTools(): Tool[] {
-  return [...TOOLS];
 }
 
 /**
@@ -322,7 +112,7 @@ export function detectTools(cwd: string = process.cwd()): DetectedTool[] {
  * Add memories MCP server to a tool's config
  */
 export async function setupMcp(
-  tool: Tool, 
+  tool: Tool,
   options: { cwd?: string; global?: boolean; dryRun?: boolean } = {}
 ): Promise<{ success: boolean; message: string; path?: string }> {
   const { cwd = process.cwd(), global: useGlobal = false, dryRun = false } = options;
@@ -334,7 +124,7 @@ export async function setupMcp(
       message: `${tool.name} does not support MCP auto-configuration`,
     };
   }
-  
+
   // Determine config path
   const resolvedMcpPath = useGlobal
     ? (tool.globalMcpConfigPath ?? tool.mcpConfigPath)
@@ -343,7 +133,7 @@ export async function setupMcp(
 
   try {
     let config: McpConfig = {};
-    
+
     // Read existing config if it exists
     if (existsSync(configPath)) {
       const content = await readFile(configPath, "utf-8");
@@ -372,10 +162,10 @@ export async function setupMcp(
       // Check if memories is already configured
       const serversKey = tool.mcpConfigFormat === "vscode" ? "servers" : "mcpServers";
       const servers = config[serversKey] ?? {};
-      
+
       if (servers.memories) {
-        return { 
-          success: true, 
+        return {
+          success: true,
           message: "MCP already configured",
           path: configPath,
         };
@@ -412,4 +202,3 @@ export async function setupMcp(
     };
   }
 }
-
