+++ packages/cli/src/lib/tool-adapters.ts
@@ -1,176 +1,27 @@
-import { writeFile, readFile, readdir, mkdir } from "node:fs/promises";
+import { writeFile, mkdir } from "node:fs/promises";
 import { existsSync } from "node:fs";
 import { join, dirname } from "node:path";
 import chalk from "chalk";
 import { MARKER, makeFooter, hasOurMarker } from "./markers.js";
 import { success as uiSuccess, warn as uiWarn, error as uiError } from "./ui.js";
+import {
+  type AdaptResult,
+  type FlatFileOptions,
+  emptyResult,
+  safeRead,
+  listMdFiles,
+  listSkillDirs,
+  safeWrite,
+  mergeSettings,
+  stripFrontmatter,
+  extractPaths,
+} from "./tool-adapters-helpers.js";
 
-// ── Types ────────────────────────────────────────────────────────────
-
-interface AdaptResult {
-  filesCreated: string[];
-  filesSkipped: string[];
-  errors: string[];
-}
-
-interface FlatFileOptions {
-  header?: string;
-  maxLength?: number;
-}
-
-// ── Helpers ──────────────────────────────────────────────────────────
-
-function emptyResult(): AdaptResult {
-  return { filesCreated: [], filesSkipped: [], errors: [] };
-}
-
-/**
- * Read a file and return its content, or null if missing/error.
- */
-async function safeRead(filePath: string): Promise<string | null> {
-  try {
-    return await readFile(filePath, "utf-8");
-  } catch {
-    return null;
-  }
-}
-
-/**
- * List .md files in a directory. Returns empty array if directory doesn't exist.
- */
-async function listMdFiles(dirPath: string): Promise<string[]> {
-  try {
-    const entries = await readdir(dirPath);
-    return entries.filter((f) => f.endsWith(".md"));
-  } catch {
-    return [];
-  }
-}
-
-/**
- * List subdirectories containing SKILL.md files.
- */
-async function listSkillDirs(skillsDir: string): Promise<string[]> {
-  try {
-    const entries = await readdir(skillsDir, { withFileTypes: true });
-    const dirs: string[] = [];
-    for (const entry of entries) {
-      if (entry.isDirectory()) {
-        const skillFile = join(skillsDir, entry.name, "SKILL.md");
-        if (existsSync(skillFile)) {
-          dirs.push(entry.name);
-        }
-      }
-    }
-    return dirs;
-  } catch {
-    return [];
-  }
-}
-
-/**
- * Write a file only if it's new or contains our marker. Returns true if written.
- */
-async function safeWrite(
-  filePath: string,
-  content: string,
-  result: AdaptResult,
-  force = false,
-): Promise<boolean> {
-  if (existsSync(filePath) && !force) {
-    const ours = await hasOurMarker(filePath);
-    if (!ours) {
-      result.filesSkipped.push(filePath);
-      return false;
-    }
-  }
-  await mkdir(dirname(filePath), { recursive: true });
-  await writeFile(filePath, content, "utf-8");
-  result.filesCreated.push(filePath);
-  return true;
-}
-
-/**
- * Deep-merge settings.json: merge allow/deny arrays, preserve user keys.
- */
-function mergeSettings(
-  existing: Record<string, unknown>,
-  incoming: Record<string, unknown>,
-): Record<string, unknown> {
-  const merged = { ...existing };
-
-  for (const [key, val] of Object.entries(incoming)) {
-    if (key === "permissions" && typeof val === "object" && val !== null) {
-      const existPerms = (merged.permissions ?? {}) as Record<string, unknown>;
-      const incomingPerms = val as Record<string, unknown>;
-
-      const mergedPerms = { ...existPerms };
-      for (const pKey of ["allow", "deny"] as const) {
-        const existArr = Array.isArray(existPerms[pKey]) ? (existPerms[pKey] as string[]) : [];
-        const incomingArr = Array.isArray(incomingPerms[pKey]) ? (incomingPerms[pKey] as string[]) : [];
-        mergedPerms[pKey] = [...new Set([...existArr, ...incomingArr])];
-      }
-      // Keep other permission keys from incoming
-      for (const [pk, pv] of Object.entries(incomingPerms)) {
-        if (pk !== "allow" && pk !== "deny" && !(pk in mergedPerms)) {
-          mergedPerms[pk] = pv;
-        }
-      }
-      merged.permissions = mergedPerms;
-    } else if (!(key in merged)) {
-      merged[key] = val;
-    }
-  }
-
-  return merged;
-}
-
-/**
- * Strip YAML frontmatter from markdown content.
- * Returns { frontmatter, body } where frontmatter is the raw YAML (without ---).
- */
-function stripFrontmatter(content: string): { frontmatter: string; body: string } {
-  const match = content.match(/^---\n([\s\S]*?)\n---\n?([\s\S]*)$/);
-  if (!match) return { frontmatter: "", body: content };
-  return { frontmatter: match[1], body: match[2] };
-}
-
-/**
- * Extract `paths:` array values from YAML frontmatter string.
- */
-function extractPaths(frontmatter: string): string[] {
-  const paths: string[] = [];
-  const lines = frontmatter.split("\n");
-  let inPaths = false;
-
-  for (const line of lines) {
-    if (line.startsWith("paths:")) {
-      inPaths = true;
-      continue;
-    }
-    if (inPaths) {
-      const match = line.match(/^\s+-\s+"?(.+?)"?\s*$/);
-      if (match) {
-        paths.push(match[1]);
-      } else {
-        inPaths = false;
-      }
-    }
-  }
-
-  return paths;
-}
+// Re-export for backward compatibility
+export type { AdaptResult, FlatFileOptions } from "./tool-adapters-helpers.js";
 
 // ── 4a: Claude Code Adapter ─────────────────────────────────────────
 
-/**
- * Adapt `.agents/` directory to Claude Code configuration.
- *
- * - `instructions.md` → `CLAUDE.md` with header and marker
- * - `rules/*.md` → `.claude/rules/*.md` (copied with marker)
- * - `skills/**​/SKILL.md` → `.claude/skills/**​/SKILL.md` (copied)
- * - `settings.json` → `.claude/settings.json` (merged)
- */
 async function adaptForClaude(
   agentsDir: string,
   outputDir: string,
@@ -195,7 +46,6 @@ async function adaptForClaude(
       const content = await safeRead(join(agentsDir, "rules", file));
       if (!content) continue;
 
-      // Ensure marker is present
       const out = content.includes(MARKER) ? content : content + makeFooter();
       await safeWrite(join(outputDir, ".claude", "rules", file), out, result);
     }
@@ -244,13 +94,6 @@ async function adaptForClaude(
 
 // ── 4b: Cursor Adapter ──────────────────────────────────────────────
 
-/**
- * Adapt `.agents/` directory to Cursor configuration.
- *
- * - `instructions.md` → `.cursor/rules/memories.mdc` with MDC frontmatter
- * - `rules/*.md` → `.cursor/rules/{name}.mdc` with paths→globs translation
- * - `skills/**​/SKILL.md` → `.cursor/skills/**​/SKILL.md` (copied)
- */
 async function adaptForCursor(
   agentsDir: string,
   outputDir: string,
@@ -261,7 +104,6 @@ async function adaptForCursor(
   try {
     const instructions = await safeRead(join(agentsDir, "instructions.md"));
     if (instructions) {
-      // Strip existing marker from source before wrapping
       const body = instructions.replace(/\n<!-- Generated by memories\.sh at .+? -->/, "");
       const frontmatter = [
         "---",
@@ -287,7 +129,6 @@ async function adaptForCursor(
       const { frontmatter, body } = stripFrontmatter(content);
       const paths = extractPaths(frontmatter);
 
-      // Build MDC frontmatter with globs
       const globs = paths.length > 0 ? paths.join(",") : "";
       const description = paths.length > 0
         ? `Rules for ${paths.join(", ")}`
@@ -301,7 +142,6 @@ async function adaptForCursor(
         "---",
       ].join("\n");
 
-      // Strip existing marker from body
       const cleanBody = body.replace(/\n<!-- Generated by memories\.sh at .+? -->/, "").trim();
       const mdcName = file.replace(/\.md$/, ".mdc");
       const out = `${mdcFrontmatter}\n\n${cleanBody}${makeFooter()}`;
@@ -330,16 +170,6 @@ async function adaptForCursor(
 
 // ── 4c: Flat-File Adapter ───────────────────────────────────────────
 
-/**
- * Adapt `.agents/` directory to a single flat markdown file.
- *
- * Concatenates instructions.md and all rules into one file.
- * Used for Copilot, Windsurf, Cline, Roo, Gemini, etc.
- *
- * @param agentsDir - Path to `.agents/` directory
- * @param outputPath - Full path to the output file (e.g. `.github/copilot-instructions.md`)
- * @param options - Optional header and maxLength for truncation
- */
 async function adaptForFlatFile(
   agentsDir: string,
   outputPath: string,
@@ -352,15 +182,12 @@ async function adaptForFlatFile(
   try {
     const sections: string[] = [];
 
-    // 1. Read instructions.md
     const instructions = await safeRead(join(agentsDir, "instructions.md"));
     if (instructions) {
-      // Strip marker from source
       const clean = instructions.replace(/\n<!-- Generated by memories\.sh at .+? -->/, "").trim();
       if (clean) sections.push(clean);
     }
 
-    // 2. Read all rules, strip frontmatter, annotate with path
     const ruleFiles = await listMdFiles(join(agentsDir, "rules"));
     for (const file of ruleFiles) {
       const content = await safeRead(join(agentsDir, "rules", file));
@@ -385,7 +212,6 @@ async function adaptForFlatFile(
 
     let content = `${header}\n\n${sections.join("\n\n")}`;
 
-    // 4. Truncate if maxLength set
     if (maxLength && content.length > maxLength) {
       const truncated = content.slice(0, maxLength);
       const lastNewline = truncated.lastIndexOf("\n");
@@ -396,7 +222,6 @@ async function adaptForFlatFile(
 
     content += makeFooter();
 
-    // Check overwrite safety
     if (existsSync(outputPath)) {
       const ours = await hasOurMarker(outputPath);
       if (!ours) {
