+++ packages/cli/src/mcp/formatters.ts
@@ -1,0 +1,45 @@
+import type { Memory, MemoryType } from "../lib/memory.js";
+
+// ─── Type Labels ──────────────────────────────────────────────────────────────
+
+export const TYPE_LABELS: Record<MemoryType, string> = {
+  rule: "📌 RULE",
+  decision: "💡 DECISION",
+  fact: "📋 FACT",
+  note: "📝 NOTE",
+  skill: "🔧 SKILL",
+};
+
+// ─── Format Helpers ───────────────────────────────────────────────────────────
+
+export function formatMemory(m: Memory): string {
+  const tags = m.tags ? ` [${m.tags}]` : "";
+  const scope = m.scope === "global" ? "G" : "P";
+  const typeLabel = TYPE_LABELS[m.type] || "📝 NOTE";
+  const paths = m.paths ? ` (paths: ${m.paths})` : "";
+  const cat = m.category ? ` {${m.category}}` : "";
+  return `${typeLabel} (${scope}) ${m.id}: ${m.content}${tags}${paths}${cat}`;
+}
+
+export function formatRulesSection(rules: Memory[]): string {
+  if (rules.length === 0) return "";
+  return `## Active Rules\n${rules.map(r => `- ${r.content}`).join("\n")}`;
+}
+
+export function formatMemoriesSection(memories: Memory[], title: string): string {
+  if (memories.length === 0) return "";
+  return `## ${title}\n${memories.map(formatMemory).join("\n")}`;
+}
+
+// ─── Tool Response Types ──────────────────────────────────────────────────────
+
+export interface ToolTextPart {
+  type: "text";
+  text: string;
+}
+
+export interface ToolResponsePayload {
+  content: ToolTextPart[];
+  isError?: boolean;
+  [key: string]: unknown;
+}
